app/requirements.txt
fastapi[standard]
pydantic
llama-cpp-python
langchain